requirements.txt
Flask==3.0.3
requests==2.31.0
pyttsx3==2.90
openpyxl==3.1.4
python-dotenv==1.0.1
